refactor: Organize project structure into src/ directory
- Created organized directory structure under src/:
  - agents/ - CLI tools and automation agents
  - core/ - Base classes and utilities
  - storage/ - Database and storage components
  - analytics/ - Analytics and reporting
  - integrations/ - External service integrations
  - validators/ - Input validation and sanitization

- Moved all Python modules from root to appropriate subdirectories
- Updated all imports throughout the codebase to use new paths
- Updated Makefile and shell scripts to use module imports
- Updated README.md to document new project structure

This improves project organization and makes the codebase more maintainable.
