+++ scripts/process_document_metrics.py
@@ -11,7 +11,7 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 import yaml
 from pathlib import Path
 from datetime import datetime
-from context_kv import ContextKV, MetricEvent
+from src.storage.context_kv import ContextKV, MetricEvent
 
 
 def process_file(file_path: str, kv: ContextKV) -> bool:
