+++ scripts/record_deployment_event.py
@@ -8,7 +8,7 @@ import sys
 import os
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from context_kv import ContextKV
+from src.storage.context_kv import ContextKV
 from datetime import datetime
