+++ src/storage/vector_db_init.py
@@ -49,7 +49,7 @@ class VectorDBInitializer:
     def connect(self) -> bool:
         """Connect to Qdrant instance"""
         # Import locally
-        from utils import get_secure_connection_config
+        from src.core.utils import get_secure_connection_config
 
         qdrant_config = get_secure_connection_config(self.config, "qdrant")
         host = qdrant_config["host"]
